requirements.txt
python-dotenv
google-api-python-client
psycopg2-binary
tqdm
requests
torch
torchvision
torchtext==0.6.0
matplotlib
numpy
scipy
spacy